{{cookiecutter.project_app_name}}/requirements.txt
alembic==1.11.3 ; python_version >= "3.11" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.11" and python_version < "4.0"
asyncpg==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.11" and python_version < "4.0"
cryptography==41.0.3 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.4.2 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.0.0.post2 ; python_version >= "3.11" and python_version < "4.0"
fastapi-pagination==0.12.8 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.101.1 ; python_version >= "3.11" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.11" and (platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64") and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and python_version < "4.0"
pydantic==1.10.12 ; python_version >= "3.11" and python_version < "4.0"
pydantic[dotenv,email]==1.10.12 ; python_version >= "3.11" and python_version < "4.0"
pyjwt[crypto]==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy2-stubs==0.0.2a35 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==1.4.41 ; python_version >= "3.11" and python_version < "4.0"
sqlmodel==0.0.8 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.11" and python_version < "4.0"
